requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
python-dotenv
requests
pandas
langchain
langchain-text-splitters